Allow for packages to be stored in a sub-directory
By default, pony-stable expects your dependency to be
stored in the top level of its repo. For example, if
you are looking for `inspect`, it expects there to be
a `inspect` directory at the top level of the GitHub
repository. Not all Pony libraries currently share that
layout.

Until we have a package manager that can unify library
directory layout, this commit add 'subdir' as a dependency
option. For example:

{
  "deps": [
    { "type": "github", "repo": "aturley/osc-pony", "subdir": "src/" }
  ]
}

which will look for any dependencies (in this case osc-pony), not
in the top level of the 'aturley/osc-pony` repo but rather in
'aturley/osc-pony/src`.