Add `version` command line option
I couldn't find a way to pass any defines to ponyc that go beyond flags
(ifdef and -D=foo don't seem to work with variables?) -- so I ended up
generating a stable/version.pony.

Issue #41 left it open how exactly this should look. This change only
outputs the version of stable,

    $ stable version
    0.1.0-af246fe [release]

and nothing related to the ponyc (and, transitively, LLVM) version used
-- does it make sense to include that? I wasn't convinced.

Thanks to @kulibali for helping a lot with replacing the string in
make.bat.