Use `fetch` instead of `pull url` to update .deps (#59)
If the `.deps` directory already existed then we were running `git -C
[root] pull [url]` and then checking out the commit specified in the
dependencies list. If the commit wasn't master then there could be a
merge conflict, which would cause the checkout to fail, leaving
`.deps` in a bad state.

The solution (for now) is to call `git -C [root] fetch` instead. This
gets the metadata without trying to merge master into the working
copy. We can then safely run the checkout command to get the latest
version of the desired dependency.