Stop looking for bundle.json if an invalid one is found
The previous behaviour was a little irritating:

If invoked in /some/nested/dir, stable would try to read

    /some/nested/dir/bundle.json

and if that failed (because the file didn't exist or was invalid), it
would try to read, subsequently,

    /some/nested/bundle.json
    /some/bundle.json
    /bundle.json

This implied that errors reading any found bundle.json were printed, but
stable kept trying to find a valid one, and eventually, having reached
/bundle.json and assuming it didn't exist, would print

    No bundle.json in current working directory or ancestors.

With this change, the logic is different: stable will first try to find
an existing bundle.json, and then attempt to read it. If it fails, it
will not go on, but stop.

Signed-off-by: Stephan Renatus <srenatus@chef.io>