Run integration tests in their own temporary directory
Having not found a way to make `@chdir[I32]` or anything like that work,
this commit adds a bash helper script that `cd`s into the directory
before running stable.

Also includes the code changes necessary to create the temporary dir
from pony, and remove it after the tests are done.

The former is done so that upcoming tests can use pony to set up some
initial situation (such as drop a bundle.json).

The cleanup is not done in a `tear_down()` method, because I found no
non-clumsy way to get the reference to the directory into that spot. So
instead, this has a mini-actor that deletes the temp dir when disposed,
and is set up to be called by the test helper.

Note that there's no strict need to run the existing integration tests
in their own temp dirs, so we might make that temp dir passing optional,
maybe, by changing `Exec`'s `cleaner arg from

    cleaner: DisposableActor

to

    cleaner: (DisposableActor | None)

and pass None for help and version.

Signed-off-by: Stephan Renatus <srenatus@chef.io>