Start integration test scaffolding with basic tests for `help` and `version` (#70)
Notably absent: setting up temp directories, doing any real changes
to the filesystem as part of the tests. This will be handled in follow-up
PRs.

Notes on this change:

* Add new make target 'integration', have 'test' exclude them

The rationale for this new target is that for #28, we'd probably have
these tests depend on git being installed; and 'make test' should still
"just work" for everyone.

Also, this is differentiating test cases by their _names_, not labels.
This is because I haven't found a way to _exclude_ based on a label, to
have `make test` not run any of the integration tests. (Short of adding
label=unit to all the unit tests, which would of course be another
option.)

* Test Usage output with different arguments to stable

I would also propose to exit with a non-zero exit code when we present
the help output because the arguments didn't match anything.

But currently, this accurately reflects the behaviour.

* Makefile: differentiate sources for $(binary) and $(test_binary)

It was a bit annoying to have

    make integration

rebuild `build/release/stable` when the only thing that had changed was
the (integration) tests.

Signed-off-by: Stephan Renatus <srenatus@chef.io>