Run integration tests sequentially
One of these tests had timed out on circleci. Having flakey tests seems
an annoyance we should try to avoid; but I couldn't reproduce this by
running `make integration` in a loop locally. So, the hypothesis is that
the timeout was no bug in the tests, but a genuine timeout.

Trying to reproduce the test failures locally, I've found that
restricting the memory to 4M (docker's minimum) caused the tests to fail
in a reproducible way.

The integration test binary is now passed `--sequential`, so that the
individual test runs don't compete for resources.

This makes the error _not appear anymore_ in the docker 4M scenario.

This reverts commit 8d8a906c69921050b18eded39159c1073afb071d.