Add integration tests for 'stable env' (#73)
* Run integration tests in their own temporary directory

Having not found a way to make `@chdir[I32]` or anything like that work,
this commit adds a bash helper script that `cd`s into the directory
before running stable.

Also includes the code changes necessary to create the temporary dir
from pony, and remove it after the tests are done.

The former is done so that upcoming tests can use pony to set up some
initial situation (such as drop a bundle.json).

The cleanup is not done in a `tear_down()` method, because I found no
non-clumsy way to get the reference to the directory into that spot. So
instead, this has a mini-actor that deletes the temp dir when disposed,
and is set up to be called by the test helper.

Note that there's no strict need to run the existing integration tests
in their own temp dirs, so we might make that temp dir passing optional,
maybe, by changing `Exec`'s `cleaner arg from

    cleaner: DisposableActor

to

    cleaner: (DisposableActor | None)

and pass None for help and version.

* Add integration tests for 'stable env'

Turns out this is probably the lowest-hanging fruit right now. This makes
use of the previous commit's "run in temp dir" feature.

* Add test case for invalid bundle.json in nested dir (valid one in parent)

This is in anticipation of what's going to change in #68.

* Run integration tests in separate tmp directory

This adds stable/test/integration/tmp/ as an empty, tracked folder; and
uses that one as base directory for running the integration tests.

It felt a bit cleaner to me than having the temp dirs created in the
top-level folder; also causes less of a mess when debugging integration
tests by disabling the cleanup dispose()'s delete call.

(I suppose it would have been just as good to use /tmp/, but I wasn't
sure what our expectations regarding different operating systems were.)

* Run integration tests sequentially

One of these tests had timed out on circleci. Having flakey tests seems
an annoyance we should try to avoid; but I couldn't reproduce this by
running `make integration` in a loop locally. So, the hypothesis is that
the timeout was no bug in the tests, but a genuine timeout.

Trying to reproduce the test failures locally, I've found that
restricting the memory to 4M (docker's minimum) caused the tests to fail
in a reproducible way.

The integration test binary is now passed `--sequential`, so that the
individual test runs don't compete for resources.

This makes the error _not appear anymore_ in the docker 4M scenario.

This reverts commit 8d8a906c69921050b18eded39159c1073afb071d.

* Double integration test timeouts

Seems like --sequential alone isn't enough. 200ms doesn't seem too bad,
so let's try this.

* integration: register tmpdir cleanup after creation
* integration/helper.sh: mention chdir issue

Signed-off-by: Stephan Renatus <srenatus@chef.io>