Initial version of AppImage packaging for pony-stable
See https://appimage.org/ for more info about the AppImage format.

The AppImage packages are built in a CentOS 7 docker image and
should theoretically be compatible with all glibc based x86_64
distros with the same or a newer glibc version.

This commit also fixes an issue with the nighly build of debian
packages.