Fix broken AppImage building
LinuxDeploy, the tool we use to build AppImage source was updated and
now creates the AppImage file with a name that is different than what
came before. This commit updates our single command that depends on that
filename to deal with the variability.

Previously files were named Pony_Dependency_Manager-x86_64.AppImage
Now they are Pony_Dependency_Manager-SOME_VERSION-x86_64.AppImage