"Remove" regex dependency
The regex library has been removed from ponyc and now lives in its
own library. This causes a bit of a problem for usage with stable.

How would we install that library so we can test stable? With the
stable we just built? Perhaps. Regex is only used in an integration
test and it can probably be accomplished in another fashion.

So that we can build stable for now and not break other things that
depend on it, I've commented out the small bit of test code that
uses regex. I will be opening an issue to come up with a fix.